requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
gunicorn
python-dotenv 

